fix: increase max variable size for th500
Increase the max variable size to prevent a case where a
revoked key couldn't be added to dbx due to max space requirement.

The ask from ARM is for a 64 KB max variable space.
For now increase this to 60KB (ARM has agreed to this) as the total
NS-S shared mbox is 64k and we can't increase it without tweaking the
StMM Manifest (doable but is more effort).

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>