feat(TegraVirt): Add android support
This change allows TegraVirt to boot Android using a boot.img loaded
into memory.  The location of the image will be passed via the DTB.

The process for this type of Android boot is very similar to an RCM
boot, so that code was commonized into a function so that it could be
reused.

Details:
- Use the NVIDIA implementation of PlatformBootManagerLib.  This allows
  us to boot directly into Android.  It requires additional libraries,
  such as PlatformBootOrderLib, FwVariableLib, IpmiBaseLib, etc. as well
  as IpmiNullDxe and BootManagerMenuApp.
- We do not implement a DevicePathProtocol, so use UefiDevicePathLib
  globally.  This is consistent with Jetson.

Signed-off-by: Jake Garver <jake@nvidia.com>