fix: Fix KSZ9031 Idle Error Bug
The KSZ9031 has a bug where it will stop working if the Idle Error
count is maxed out. This can occur on power up on systems without
ethernet magnetics and requires a soft reset if it occurs.

This patch adds an Idle Error check and soft reset based on the Linux
driver.

Signed-off-by: Parker Newman <pnewman@connecttech.com>