feat: Add UEFI Adapter Information Protocol Support
This adds support for the UEFI Adapter Information Protocol (AIP) for Nvidia
Ethernet controllers. This fixes an issue where PXE does not load if
the network interface is not ready (MediaPresent) when the PXE driver starts.

Without AIP support NetLibDetectMediaWaitTimeout() will check if the network
interface is ready only once. If it is not then it cancels PXE boot. This
means if the network interface is slow to link the system will not attempt
to PXE boot.

When AIP support is availible NetLibDetectMediaWaitTimeout() will poll for
"MediaPresent" to be "TRUE" or until it times out.

Based on: edk2-platforms/Silicon/Broadcom/Drivers/Net/BcmGenetDxe/
Found at: https://github.com/tianocore/edk2-platforms/

Signed-off-by: Parker Newman <pnewman@connecttech.com>