fix: Tpm command returns buffer too small
The TPM chips that were originally tested with returned burst count as
the number of bytes that the last TPM command responded with. Hence, the
issue was not seen.

With some other TPM chips, the burst count was the FIFO size, which was
larger than the expected response size, hence the code returned
BUFFER_TOO_SMALL.

This patch fixes the issue by receiving based on the expected response
size instead of burst count.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>