fix(android): update logic for new device hierarchy
Currently, if requested, we add new devices to top of boot order only if
the device is a removable device. However, this may not be desirable for
android use cases where different boot modes would be there from the
same media. In this change, detect android boot and if set, skip
checking for removable device while setting boot order.

Signed-off-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Bob Morgan <bobm@nvidia.com>
Tested-by: Yijun Zhou <yijunz@nvidia.com>