feat: Add QSPI combined sequence mode support
This patch adds an option for callers to perform QSPI transactions in
combined sequence mode, in which the command and address are explicitly
specified.

By default, the combined sequence mode is disabled. Hence, the other
drivers that have been using QSPI driver and been passing
command/address over TxBuf will not need any updates, and will still work
as is.

Signed-off-by: Dat Mach <dmach@nvidia.com>