fix: hang due to no QSPI timeout if TPM not present
QSPI timeout was changed to infinite loop to workaround some issue in
the past. And since TPM wait state was added, QSPI controller would wait
for TPM to get out of wait state. Because of the above, it would hang
polling QSPI ready bit forever if TPM chip is not present.

This patch adds QSPI timeout back to fix the issue and increases the
timeout value to 1 second to maintain the old workaround.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>