fix: SW_IO_4 usage for DPC handling
As part of the DPC handling when the OS calls the _OST() method
of the root port, it is expected to raise an interrupt to the RAS-FW
through SW_IO_4 of Socket-0 always. Because of an unintended change,
currently, SW_IO_4 of respective socket to which the RP belongs, is
raised to the RAS-FW because of which the DPC handling flow is broken.
This change reverts it back to using SW_IO_4 of Socket-0 always.

Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>