chore: update DTB cache parsers to use new format
DeviceTreeHelperLib now has a function for parsing CacheData from
the device tree, so use that when generating the CacheInfo in Pptt
instead of parsing the device tree information directly.

Also checks the return code for cache parsing, prints errors, and
deduplicates code.

Suppport updated DTB paths in FloorSweepingLib, and updated properties
in APMT.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Tested-by: Jeff Brasen <jbrasen@nvidia.com>