feat: Add driver for uploading Renesas USB FW
On Grace platforms, Renesas USB Controller needs to have an external
SPI ROM to carry its firmware. However, some customers may want to
remove the SPI ROM for cost and/or security reason.

This patch allows UEFI to carry the uPD720201 firmware and load it after
PCI enumeration so that the USB controller will work on platforms that
do not have the external SPI ROM installed.

The firmware binary can be downloaded from Renesas website and store in
UEFI FV as gNVIDIAuPD72020xFirmwareGuid.

This driver relies on UEFI variable "LoadRenesasUsbFw" to know whether
the platform needs the USB firmware loaded or not. By default, this
variable is not set, hence, this code doesn't run for most platforms.
Only the platform that doesn't have the SPI ROM would need to create
this variable in device tree overlay, for the code to run.

The driver also dynamically checks if the SPI ROM exists. And when it
does, it would exit without overwriting the FW regardless of whether the
variable "LoadRenesasUsbFw" is set or not.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>