fix: Improve UPHY configuration flexibility
Currently, UPHY2 and UPHY3 are only allowed to configure to NVLINK and
not PCIe when C2C_MODE is 2. Going forward, we'll need to improve
flexibility.

This patch does not use C2C_MODE but honors the current boot UPHY
setting instead. If UPHY2/3 had been set to NVLINK at boot then block
PCI settings for it, and vice versa.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Simon Wang (SW-GPU) <simowang@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>