feat: hook to change CPER destinations
When RAS_FW calls the UEFI MM API to store a CPER in flash, provide a
hook to allow setting/clearing some of the possible destinations (HEST,
BMC). This allows a customer to make changes based on their
requirements. An example is provided to prevent any corrected or
informational CPER from being sent to HEST.

Signed-off-by: nbenech <nbenech@nvidia.com>
Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>