fix: protect UPHY2/3 from switching PCIE to NVLINK
Mercury modules have UPHY2/3 preset to NVLINK and these settings are
stored in UEFI variables. When booting a platform that has UPHY2 set as
PCIE, it gets overwritten with the values in UEFI variables and UPHY2
gets reset to NVLINK, which will cause the hard lock in MB1 on
successive boots.

This patch adds a check to honor the current boot UPHY2/3 settings and
only apply the UEFI variable values if they are of the same types.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>