feat: Increase LZMA dictionary size
Increase the LZMA dictionary size to improve compression ratio in the
image.

This requires a change in tools_def.txt since this file is parsed
directly by GenFdsGlobalVariable.py.  Rather than fork the template and
maintain our own tools_def.txt, we'll an NVIDIA-specific include at the
end of the file.

Going forward, we'll be able to override or add tools_def.txt directives
with changes to Silicon/NVIDIA/tools_def.inc.

Signed-off-by: Jeff Brasen <jbrasen@nvidia.com>
Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>