fix: Enable DPC SIG_SFW always
Keep DPC SIG_SFW enabled always irrespective of the DPC enablement.
In the case of RP, this makes sure that the DPC gets triggered at the
RP level upon receiving an ERR_COR message with its sub-class set to
ECS SIG_SFW which is required to support Hot-Plug for the devices
connected behind a PCIe switch.

Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>