feat: Add CM API for adding an AcpiTableGenerator
This adds a new API function that will add the passed in ACPI table
generator to the ACPI table generator list if that generator isn't
already present in the list. If it is present, it will print warning
messages about any incompatibilities and not add the table.

The AcpiTableListParser is also updated to use this API.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>