fix: Strip git output when generating firmware version
The command "git describe" will output a terminating linefeed.  That
hasn't caused a problem because the BUILDID_STRING is the last thign on
the build command.  However, if an additional build variable is set, the
terminating linefeed pushes the new build variable to a new line.

This fix adds a .strip() to git's output.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>