chore(apei): send C2C GPU flag to RAS_FW
On some platforms, it is possible to create a DTB overlay that
explicitly disables C2C for PCIe controllers (for example in the case of
a C1 platform). This patch adds a call into RAS_FW to indicate that
fact, which then allows RAS_FW to enable/disable some corresponding
features.

Signed-off-by: Nicolas Benech <nbenech@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>