fix(stmm): allow measurement partition to be zero filled
Currently if the measurement partition is erased then we force a re-init
of the variable partition. But depending on how the device is
provisioned it is possible that the partition is zero filled, make a
code change to treat this similar to an erased partition.

Provide an additional option on what action is to be taken if the
Variable store integrity check fails. Currently, the default behavior is
to assert.

Add configurability to either assert or re-initialize/reboot if the
Variable store integrity has failed.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>