fix: IORT assert should check NumberOfNodes
The IORT code had a code sanity check ASSERT that tried to make sure the
loop counter didn't exceed the number of nodes, but the ASSERT was
incorrectly comparing to the SizeOfNode field rather than the
NumberOfNodes field.

On systems with a large number of nodes this could result in the ASSERT
incorrectly being hit.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>