fix: Avoid overrunning GOP screen when printing version
When printing the system firmware banner, we need to limit it to the
width of the screen.  Printing longer strings could overrun the buffer
on some GOP drivers.  This was an issue when BUILDID_STRING was too
long, but could also be a problem if a longer platform name was used.

To avoid overrunning, we'll now check the screen width and clip the
system firmware banner appropriately.

Also, to avoid clipping the banner, which is likely in low resolutions,
we'll split the banner across three lines.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Aron Wong <awong@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>