feat: DTB helper defaults for cell sizes
The DTB spec (v0.4) says to use default values of 2 and 1
for #address-cells and #size-cells, respectively, if the properties are
missing.

This updates the DeviceTreeGetNodePropertyValue64 function to check if
those properties are being queried and return defaults instead of
EFI_NOT_FOUND when they are missing. Note that the 32-bit function wraps
the 64-bit function, so also gets this change as a result.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>