fix: reset the meas buffer after computing the first measurement
After using the measurement buffer to verify the variable
store, reset it back to 0.
Without this change, if a var write comes in the FVB driver
treats the contents of the measurement buffer as valid to be
committed and it throws the state machine off.
Also return an error when checking if the Measurement partition
is zero or erased to force a re-init of the VarStore.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>