feat(build): Convert an empty config var to an undefined config var
Kconfig allows a variable to be empty.  But edk2's build system does
not.  As part of our config file post-processing, convert empty
variables to undefined variables.

This is convenient when it comes to optional config strings.  With this change
in place, the strings can default to empty.  Without this change, we'd
need to have a boolean guard around it in the menuconfig.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>