feat(build): Support a LOCALNAME definition in Kconfig
This concept is taken from the linux kernel, which uses LOCALVERSION as
suffix to the upstream version.  Here, the LOCALNAME will be appeneded
to BUILD_NAME to create the PLATFORM_NAME, allowing derivatives of
upstream platforms to be easily identified.

The OUTPUT_DIRECTORY was redefined to use BUILD_NAME instead of
PLATFORM_NAME to work-around a limitation in stuart.  The stuart build
wrapper ignores conditional directives when parsing DSC files.  It is
only interested in a few variables, so this isn't usually a problem.
However, OUTPUT_DIRECTORY is one of those variables.  Using BUILD_NAME
insted of PLATFORM_NAME means stuart and build.py will always agree on
OUTPUT_DIRECTORY.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>