chore: Clean up CM depex to make sure it runs
Previously if certain entries in DTB were disabled
it could result in protocols related to those not
being installed, which would then prevent
ConfigurationManagerDataDxe from running if it had
a depex on those protocols. This changes the depex
to simply be one that gets created late enough in
the boot to be sure that all the things CM wanted
to run after have already had their chance to run.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>