refactor: Allow unittests to be built with --whole-archive
The UnitTestFramework now links host binaries with --whole-archive,
which means unused portions of libs are no longer optimized out at link
time.

This change means we can no longer include the code under test as a NULL
library.  This was convenient, but results in undefined entry point
symbols.  Instead, we'll need to include the code under test as Sources
in the unittest binary.

This change exposed a few additional build issues:
* Missing declarations.  Resolved by including missing headers.
* Unnecessary libraries.  Resolved by removing them from unittest
  binaries.
* Multiple definitions of a global variable, Channel.  Resolved by
  renaming the test's Channel to TestChannel.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>