feat: add new ns mailbox for a PRM service
Change that adds a new NS-S Mailbox. A UEFI driver needs to make an FF-A
call to get the PA/Size of this mailbox, similar to what the
MmCommunicate Driver does in GetBufferAddr(), except you pass in
STMM_GET_PRM0_BUFFER for Arg3 (STMM_GET_PRM0_BUFFER isn't in a common
header at the moment, so may need to redefine this, there aren't too
many shared secure/non-secure headers).

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>