chore: Clean up CacheId for MPAM requirements
MPAM requires the L3 cache to have a CacheId that is only 8 bits, so
renumber the CacheInfo->Type enums to make that happen when the ID is
generated.

The GetCacheId function is also added to replace the macro that
generates the ID. This allows the function to do sanity checking via
ASSERTs for all invalid conditions.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>