feat: Run DsdtPatcher for all chip types
DsdtPatcher was only being run for server chip types, but now there is a
need to run it on a non-server chip. This changes it to run on all chip
types, and to be non-fatal for EFI_NOT_FOUND errors in patching (which
will happen on chips that don't need that field to be patched).

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Bob Morgan <bobm@nvidia.com>