fix: fail instead of ASSERT on bad reg DTB
Previously if DeviceTreeGetRegisters couldn't find the #address-cells
or #size-cells fields it would return an error. Another commit is
updating the code to use default cell values, which can result in an
ASSERT being hit if the defaults don't match the reg property's size.
This fix changes the failing ASSERT to instead print an error message
and return an error code, similar to what would previously happen if the
fields weren't found.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Bob Morgan <bobm@nvidia.com>