fix: Add DeInit in UsbPadCtl in T194
Need to set ELPG_CLAMP_EN in ELPG_PROGRAM_1 for SS part when exiting to kernel.
Otherwise, there is a chance that the SS portion cannot be enabled in kernel.

Also remove the T194-xudc in XusbControllerDxe.c to avoid PG on/off.
We do not need T194-xudc in UEFI.

Signed-off-by: Rick Tseng <rtseng@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>