fix: DTB ranges parsing with DeviceTreeHelperLib
The DeviceTreeGetRanges function is updated to get the proper
address-cells and size-cells parameters from the correct locations, per
the DTB specification. The ChildAddress field is enhanced to support
3 cells, which is used by the PCIE controller.

The PCIE controller is updated to use the fixed function for "ranges"
and "hbm-ranges".

The IORT parser is updated to use the fixed function for "dma-ranges".

UnitTests are updated to support these chagnes.

Signed-off-by: Jeshua Smith jeshuas@nvidia.com
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Tested-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jeshua Smith <jeshuas@nvidia.com>