feat: Build types in Kconfig
This rearchitects the Kconfig system to be based on selecting a build
type and an SOC as a starting point and then making tweaks to the
selected options from there. Dependency handling is also improved.

Specifically:
- SOC type will list which hardware-specific features are supported on
that SOC, and will automatically disable features that the hardware
cannot support.
- Build types will imply that certain features or groups of features are
enabled by default, but can be overridden if the user wants to.
- Dependencies for a number of features have been specified using a
combination of imply and default statements that attempt to auto-enable
dependencies and with comments that will flag invalid configs for unmet
dependencies.

Note: With this change only a single SOC can be selected for the build.
This is a simplification choice that helps prevent creating a build that
tries to enable incompatible features. If the need for multiple SOCs in
a single build arises this choice can be reconsidered.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>