fix: Incorrect Payload size of a write command to MB1 causes failure
Within AccessMb1Record, the payload size used for reading and
writing the "EarlyVariable->Data.Mb1Data" structure is set to
sizeof (Payload->Data) which is equal to 2040. This is larger
than the actual size of the structure which is equal to 1472
(sizeof (TH500_MB1_CONFIGURATION)). During the read operation
there is no issue since the EarlyVariable structure is allocated
with an extra buffer, but during write operation we are sending
extra data beyond the size of the structure and that is causing
an overflow which leads to a system to crash. With this fix the
size of the payload is set to sizeof (TH500_MB1_CONFIGURATION).

Signed-off-by: Kashif Memon <kmemon@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>