refactor: Remove ResetLib dependency from BaseDebugLibSerialPort
In BaseDebugLibSerialPort, avoid depending on ResetLib by replacing the
call to ResetCold() with a private function.

We have been using a deprecated ResetLib implementation, which has been
removed in edk2-stable202408.  The replacement ResetLib depends on
DebugLib, which puts us in a circular dependency.  Break the circle by
removing our DebugLib's dependency on ResetLib.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>