fix: Allow NULL ElementTokenMaps
ElementTokenMaps don't make sense for EArmObjCmRef objects, and the code
that tried to limit them to one element maps mismatched with the Find
function resulting in the Find function searching data outside of the
map for matching tokens. Instead, change the code to allow NULL
ElementTokenMaps and use a NULL map for EArmObjCmRef.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>