fix: Move rela.dyn outside of .text to satisfy llvm linker
This resolves a section type mismatch for .rela.dyn and .rela.plt when
linking with the llvm linker.

    ld.lld: error: section type mismatch for .rela.dyn
    >>> <internal>:(.rela.dyn): SHT_RELA
    >>> output section .text: SHT_PROGBITS

    ld.lld: error: section type mismatch for .rela.plt
    >>> <internal>:(.rela.plt): SHT_RELA
    >>> output section .text: SHT_PROGBITS

The mismatch is avoided by moving .rela.dyn and .rela.plt out of .text.
This allows us to continue using clang for static analysis.

Unfortunately, this results in an unbootable image when built with
clang.  A follow-on change will address the boot issue.

Signed-off-by: Jake Garver <jake@nvidia.com>