feat: Generate defconfig
When building the config file, also generate a defconfig as a
convenience.

A defconfig is a minimal config file.  Instead of recording the value of
every config variable, it omits values that can be derived from Kconfig.
As such, it's useful in providing a concise view of the image's
configuration, but without the matching Kconfig file, it does not
provide a complete view.

Also, ensure .config and defconfig are consistently generated with
DOS-style line endings.  This aids consistency when these files are put
under source control.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>