fix: Leave DisplayHandoff enabled on ACPI boot
This change leaves the DisplayHandoff enabled on ACPI boot. This is
necessary to ensure that the clocks are still enabled when in ACPI mode
as the OS does not have control of these.

Also, disable the menu if the NV GOP driver is not part of the build or
if ACPI mode is enabled.

Signed-off-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>