fix: Type 19 range size when EGM enabled
Currently, Type 19 parser uses socket number to index DramRegions. It
works when EGM is disabled, in which case the number of DramRegions
equals to the number of sockets.

However, when EGM is enabled, UEFI carveout will also be added right
after DramRegion of socket 0. This will shift DramRegions of other
sockets causing incorrect Type 19 settings.

This patch reorders EGM related carveouts to after the end of
DramRegions to fix the issue.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>