feat: Dynamic condition for Type 9 SMBIOS overlay
Currently, SMBIOS Type 9 contains static slot info of a certain PCIe
topology. However, some platforms could have multiple topologies that
are slightly different from the others. It is desired to be able use a
single SBIOS that can dynamically detect and update Type 9 tables.

This patch adds code to process the 'condition' subnode to check and
decide whether to install Type 9 or not.

Examples:

    type9@1 {
      condition {
        type = "pcie";
        address = <PCIE_SBDF(9,1,0,0) 0x00>;
        value = <0x102315B3 0xFFFFFFFF>;
      };
      ...
    };

    type39@0 {
      condition {
        type = "fru-desc";
        value = "PDU_FRU0";
      };
      ...
    };

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Simon.WJ Wang (SW-GPU) <simowang@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>