feat: BootConfigProtocolLib
This library provides an API for a BootConfigUpdateProtocol which can
be used to store up additions to make to bootconfig. If code calls the
APIs for getting the protolol or for adding the SN to the protocol and
it hasn't been installed yet, this API will install it before using it.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>