fix: memory corruption in RedfishChassisInfoCollectorDxe
The variable 'Length' of type INT32 is being typecasted to UINTN pointer
and passed in to gRT->GetVariable. Since Length is only 4 bytes but
gRT->GetVariable writes an 8 byte value to it, this would clear the next
4 bytes on stack, which coincides to be 'DeviceTreeBase' pointer.

In normal circumstance, the corrupted DeviceTreeBase pointer may go
unnoticed. But when EGM with 256MB HV enabled, the corrupted pointer
points to reserved address and triggers exception.

This patch allocates new variable of type UINTN to use in GetVariable to
fix this issue.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>