feat(build): Implement --skip-verify
This option will allow dependency verification to be skipped.  This is a
replacement for "--noverify", which has since been removed from
stuart_build.

The "--skip-verify" option works by removing all scopes.  The
verification step of stuart is still run, but it becomes a no-op.

A different name for the option was used to avoid confusion.

BREAKING CHANGE: Subclasses of NVIDIASettingsManager should no longer
override GetActiveScopes() to add scopes.  Instead, they should override
GetAdditionalScopes().

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeshua Smith <jeshuas@nvidia.com>