fix(stmm): Avoid installing the system memory HOB
During the StandaloneMmEntryPoint we install resource HOBs that the
StandaloneMm Core can then use.
A new change to the Core code checks that the mmram ranges don't
overlap with System Memory HOBs. To avoid failing this check
don't install the System Memory HOB since we already describe
all the DRAM ranges (allocated or not) when setting up the
mmram ranges part of the gEfiMmPeiMmramMemoryReserveGuid HOB.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>