feat: Add scaled logo support
Add support for scaled logos. This will take the specified logo and
scale it to the screen size.

Currently the logo is scaled to 40% of the screen height and centered
vertically. Places the center of the logo at 38.2% of the screen height.

Introduced Kconfig options to control the screen ratio and center y
position of the logo. If scaling ratio is set to 0, the logo will not be
scaled.

Added a single black logo in addition to the multiple gray logos.

Moved the logo source files to the driver to allow for easier control
with Kconfig options.

REF: https://learn.microsoft.com/en-us/windows-hardware/drivers/bringup/boot-screen-components
Signed-off-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>