feat(android): Migrate DTBs to a DTBO partition
Google has mandated that Android needs to have dtbo partition.
https://source.android.com/docs/core/architecture/dto/partitions

Google mandates that dtbo partition needs to contain the dtbo(s) in a
specific format. All dtbos of a specific SoC are bundled into a single
dtbo.img, the bootloader is supposed to apply overlays to the base dtb
at runtime.

Signed-off-by: Frankie Chang <frankiec@nvidia.com>
Reviewed-by: Jian-Min Liu <jianminl@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>