fix: get the device cid from the diskinfo protocol
Get the Device's CID register value from the Diskinfo Protocol
that is installed for other partitions instead of sending
CMD10 to get the contents of the value as the trying to
read the register contents using CMD10 seems to fail once
the initial boot config is done.
The DiskInfo protocol is not installed for the RPMB partition
so look this up indirectly via BlockIO protocol installed on
other partitions.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>