fix: methods to select rpmb partition and get emmc device
Fix the command that sets/clears the RPMB the partition.
Fix the logic that finds the right eMMC controller by
setting the slot to 0xFF before looking for slots on
each pass through device controller.
Add the documentation for the functions.
Remove unused functions such as methods that request status
from the device.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>