Fix: Cancel FW update on all FDs if an error occurs on any FD
Currently, when an update fails on one Erot due to errors, the update
continues on other Erot devices and eventually exits with an error
status. If the system resets at this stage, the board fails to boot due
to a partial firmware update, resulting in cross-socket check errors.

With these changes, various stages of the state machine are updated to
transition to the relevant cancel state on all Erots if any of the Erots
fail.

Signed-off-by: Thejaswani Putta <tputta@nvidia.com>
Reviewed-by: Bob Morgan <bobm@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>