feat(NvDisplayControllerDxe): introduce 'efifb' display hand-off method
Introduce configurable 'efifb' SOC display hand-off method. When this
method is active, we update the kernel boot command line in case the
Linux 'efifb' kernel driver can be used, i.e. when there is an active
child GOP instance with a suitable framebuffer.

The additional kernel arguments prevent the kernel from cutting power
and clocks to the display, since it currently has no way of knowing
the display is backing the EFI framebuffer.

Signed-off-by: Jan Bobek <jbobek@nvidia.com>