fix: SSIF force timeout window for faster recovery
Currently the code checks the last 10 SSIF transactions and forces error
if the average transaction time is longer than 1 second. This causes the
system to keep forcing errors even when transaction times have returned
to normal after peaking.

This change reduces the window size to 4 transactions in 12 second to
make it more reactive to transaction time changes.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Tested-by: svc-sw-mobile-l4t <svc-sw-mobile-l4t@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>