fix(qspi): adjust check for reset state
Currently, the code checks if the controller is in a reset state by
checking if the command register matches the reset value. However, in
some chips post suspend exit, only some bits are reset, so check if the
bits the code initializes are reset to decide the state of the
controller.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>