fix(stmm): Avoid MmPrepareCommunicationBuffer
A new change to the StandaloneMmCore looks for a new HOB type
that describes the MM communication buffer gMmCommBufferHobGuid
and sets up a shadow for this buffer.

However in our deployments we use the gEfiStandaloneMmNonSecureBufferGuid
HOB to describe the MM communication buffer and the Cpu driver
sets up a shadow buffer before routing the MMI.

Comment out the step to setup a shadow in the MM Core for now.
The current updates seem extremely x86 centric , we need time to explore
this for FF-A deployments.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>