fix(build): Avoid double tag name with annotated tags
Prior to this change, if we ran into an annotated tag, we'd duplicate
the tag name, resulting in a BUILDID_STRING like
"tagname-tagname-3-g4ba123ab9".  After this change, we'll report the
tagname once, plus the SHA (for example, "tagname-4ba123ab9").  This
will be the default, but can still be overridden by environment
variables.

Signed-off-by: Jake Garver <jake@nvidia.com>